+++ Magic/CHANGELOG.md
@@ -22,6 +22,7 @@
  - Add `critical:` effects section for Damage action, when a critical hit occurs
  - Add more custom wand lore placeholders: $attributes, $slots, $slot, $sets, $potion_effects, $enchantments, $cast_spell
  - Add zAuctionHouse support, bound wands can't be auctioned
+ - Add `placeholders` option to wands, for making custom placeholders that can be used in custom lore
 
 ## Balance Changes